Support for Mill 0.11.0-M11
Updated Mill API version and Scala versions
