+++ build.sc
@@ -17,7 +17,7 @@ val rtMillVersion = build.version
 trait Deps {
   def millPlatform: String
   def millVersion: String
-  def scalaVersion: String
+  def scalaVersion: String = "2.13.11"
   def itestVersions: Seq[String]
 
   val aspectjTools = ivy"org.aspectj:aspectjtools:1.9.5"
@@ -32,32 +32,28 @@ trait Deps {
 
 object Deps_0_11 extends Deps {
   override def millPlatform = millVersion // only valid for exact milestone versions
-  override def millVersion = "0.11.0-M7" // select only exact milestones
-  override def scalaVersion = "2.13.10"
+  override def millVersion = "0.11.0-M11" // select only exact milestones
   override def itestVersions = Seq(millVersion)
 }
 object Deps_0_10 extends Deps {
   override def millPlatform = "0.10"
   override def millVersion = "0.10.0" // scala-steward:off
-  override def scalaVersion = "2.13.10"
   override def itestVersions = Seq("0.10.11", millVersion)
 }
 object Deps_0_9 extends Deps {
   override def millPlatform = "0.9"
   override def millVersion = "0.9.3" // scala-steward:off
-  override def scalaVersion = "2.13.10"
   override def itestVersions = Seq("0.9.12", millVersion)
 }
 object Deps_0_7 extends Deps {
   override def millPlatform = "0.7"
   override def millVersion = "0.7.0" // scala-steward:offf
-  override def scalaVersion = "2.13.10"
   override def itestVersions = Seq("0.8.0", "0.7.3", millVersion)
 }
 object Deps_0_6 extends Deps {
   override def millPlatform = "0.6"
   override def millVersion = "0.6.0" // scala-steward:off
-  override def scalaVersion = "2.12.17"
+  override def scalaVersion = "2.12.18"
   override def itestVersions = Seq("0.6.3", millVersion)
 }
